Implemented Potions
Potions can be dropped, enemies can take damage. Damage indicator for enemies also included...

+++ Assets/Alchemy.cs
@@ -17,7 +17,7 @@ namespace Alchemy {
 		[SerializeField]
 		public int size { get; set; }
 		[SerializeField]
-		public int time { get; set; }
+		public float time { get; set; }
 		[SerializeField]
 		public int damage { get; set; }
 		[SerializeField]
